chore: factorise same functions in api
