+++ front/app/api/cron/crank/route.ts
@@ -1,47 +1,16 @@
 import { NextRequest } from "next/server";
 import { randomUUID } from 'crypto';
 import * as anchor from "@coral-xyz/anchor";
-import { PublicKey, Connection, Keypair, SystemProgram, Transaction } from "@solana/web3.js";
-import { Program, AnchorProvider } from "@coral-xyz/anchor";
-import fs from "fs";
-import path from "path";
-import { fileURLToPath } from 'url';
-
-// Polyfill pour __dirname en ESM
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
-
-// Importer l'IDL directement 
-import idlJson from "../shared/programs.json";
-
-// Configuration simple
-const RPC_ENDPOINT = process.env.SOLANA_RPC_ENDPOINT || "https://api.devnet.solana.com";
-
-// Note: La structure de l'IDL peut différer
-const idlAddress = "address" in idlJson 
-  ? (idlJson as any).address
-  : (idlJson as any).metadata?.address || "3HBzNutk8DrRfffCS74S55adJAjgY8NHrWXgRtABaSbF";
-
-// Récupération de la seed admin depuis les variables d'environnement
-const ADMIN_SEED_BASE64 = process.env.ADMIN_SEED_BASE64;
-
-console.log("📝 Adresse du programme trouvée:", idlAddress);
-const PROGRAM_ID = new PublicKey(idlAddress);
-
-// Préparer l'IDL pour Anchor - identique à program.ts
-const idl = {
-  ...idlJson as any,
-  address: idlAddress // S'assurer que l'adresse est toujours disponible à la racine
-};
-
-console.log("📝 IDL Details:", {
-  name: idl.metadata?.name || "Unknown",
-  version: idl.metadata?.version || "Unknown",
-  address: idl.address,
-});
-
-// Log des instructions disponibles
-console.log("📝 Instructions disponibles:", idl.instructions?.map((ix: any) => ix.name) || []);
+import { PublicKey, Connection } from "@solana/web3.js";
+import { 
+  getAdminKeypair, 
+  getProgram, 
+  createAnchorWallet, 
+  verifyAuthToken, 
+  RPC_ENDPOINT,
+  createSuccessResponse,
+  createErrorResponse
+} from "../shared/utils";
 
 // --- Définition des interfaces ---
 interface EpochManagementAccountInfo {
@@ -84,83 +53,6 @@ interface EpochAccount {
     processed: boolean;
 }
 
-// Définir l'interface pour le wallet Anchor
-interface AnchorWallet {
-  publicKey: PublicKey;
-  signTransaction: <T>(tx: T) => Promise<T>;
-  signAllTransactions: <T>(txs: T[]) => Promise<T[]>;
-}
-
-// Fonction pour vérifier le token d'authentification
-function verifyAuthToken(request: NextRequest): boolean {
-  const authHeader = request.headers.get('authorization');
-  if (!authHeader) {
-    console.error('❌ En-tête Authorization manquant');
-    return false;
-  }
-
-  const token = authHeader.replace('Bearer ', '');
-  const expectedToken = process.env.API_SECRET_KEY;
-
-  if (!expectedToken) {
-    console.error('❌ API_SECRET_KEY non défini dans les variables d\'environnement');
-    return false;
-  }
-
-  return token === expectedToken;
-}
-
-// Fonction pour générer le keypair admin à partir de la seed stockée en Base64
-function getAdminKeypair(): Keypair {
-  if (!ADMIN_SEED_BASE64) {
-    throw new Error("ADMIN_SEED_BASE64 n'est pas défini dans les variables d'environnement ou est vide");
-  }
-  
-  try {
-    const seedBuffer = Buffer.from(ADMIN_SEED_BASE64, 'base64');
-    if (seedBuffer.length === 32) {
-        console.log("✅ Utilisation de Keypair.fromSeed() car seedBuffer fait 32 bytes.");
-        return Keypair.fromSeed(seedBuffer);
-    } else if (seedBuffer.length === 64) {
-        console.log("✅ Utilisation de Keypair.fromSecretKey() car seedBuffer fait 64 bytes.");
-        return Keypair.fromSecretKey(seedBuffer);
-    } else {
-        console.error(`❌ Taille de seedBuffer inattendue: ${seedBuffer.length} bytes. Devrait être 32 ou 64.`);
-        throw new Error(`Taille de seedBuffer après décodage Base64 inattendue: ${seedBuffer.length}`);
-    }
-  } catch (error) {
-    console.error("❌ Erreur lors de la génération du keypair admin (détail):", error instanceof Error ? error.message : String(error));
-    throw new Error(`Impossible de générer le keypair admin: ${error instanceof Error ? error.message : String(error)}`);
-  }
-}
-
-// Fonction pour obtenir le programme Anchor - Identique à program.ts
-function getProgram(connection: Connection, wallet?: AnchorWallet | null) {
-  try {
-    // Création du provider avec le wallet et la connexion
-    const provider = new AnchorProvider(
-      connection,
-      wallet ?? ({} as AnchorWallet), // allow "read-only" mode comme dans program.ts
-      { preflightCommitment: "processed" }
-    );
-
-    if (wallet) {
-      console.log("⚙️ Création du provider avec wallet:", wallet.publicKey.toString());
-    } else {
-      console.log("⚙️ Création du provider en mode lecture seule");
-    }
-    
-    // Création du programme - EXACTEMENT comme dans program.ts
-    console.log("⚙️ Création du programme avec IDL complet...");
-    const program = new Program(idl as any, provider);
-    
-    return program;
-  } catch (error) {
-    console.error("❌ Error creating program:", error);
-    return null;
-  }
-}
-
 /**
  * Exécute la logique principale du Crank.
  * Charge la configuration depuis les variables d'environnement.
@@ -185,23 +77,7 @@ async function runCrankLogic(): Promise<{ success: boolean; message: string; det
         }
         
         // Création du wallet avec le keypair admin
-        const wallet: AnchorWallet = {
-            publicKey: adminKeypair.publicKey,
-            signTransaction: async <T>(tx: T): Promise<T> => {
-                if (tx instanceof Transaction) {
-                    tx.partialSign(adminKeypair);
-                }
-                return tx;
-            },
-            signAllTransactions: async <T>(txs: T[]): Promise<T[]> => {
-                return txs.map(tx => {
-                    if (tx instanceof Transaction) {
-                        tx.partialSign(adminKeypair);
-                    }
-                    return tx;
-                });
-            },
-        };
+        const wallet = createAnchorWallet(adminKeypair);
         
         const program = getProgram(connection, wallet);
         
@@ -296,7 +172,7 @@ async function findClosedUntreatedEpochs(program: anchor.Program<any>): Promise<
  */
 async function processEpoch(
     program: anchor.Program<any>,
-    adminAuthority: Keypair,
+    adminAuthority: anchor.web3.Keypair,
     epochPda: PublicKey,
     epochAccount: EpochAccount
 ): Promise<{ success: boolean; message: string}> {
@@ -436,16 +312,10 @@ export async function GET(request: NextRequest): Promise<Response> {
   // Vérification du token d'authentification
   if (!verifyAuthToken(request)) {
     console.error(`[${requestId}] ❌ Authentification échouée`);
-    return new Response(JSON.stringify({
-      success: false,
-      error: 'Non autorisé',
-      errorType: 'AuthenticationError',
-      timestamp: new Date().toISOString(),
-      requestId,
-    }), {
-      status: 401,
-      headers: { 'Content-Type': 'application/json' },
-    });
+    return createErrorResponse(requestId, {
+      message: "Non autorisé",
+      name: "AuthenticationError"
+    }, 401);
   }
 
   try {
@@ -463,32 +333,8 @@ export async function GET(request: NextRequest): Promise<Response> {
       console.log(`[${requestId}] ⚠️ Le crank a rencontré un problème: ${result.message}`);
     }
     
-    return new Response(JSON.stringify({
-      ...result,
-      timestamp: new Date().toISOString(),
-      environment: process.env.NODE_ENV,
-      requestId,
-    }), {
-      status: 200,
-      headers: { 'Content-Type': 'application/json' },
-    });
+    return createSuccessResponse(requestId, result);
   } catch (error) {
-    const errorType = error instanceof Error ? error.name : typeof error;
-    const errorMsg = error instanceof Error ? error.message : String(error);
-    const errorStack = error instanceof Error ? error.stack : undefined;
-    console.error(`[${requestId}] ❌ Erreur lors de l'exécution du crank:`, errorMsg, errorStack);
-    
-    return new Response(JSON.stringify({
-      success: false,
-      error: errorMsg,
-      errorType,
-      stack: errorStack,
-      timestamp: new Date().toISOString(),
-      environment: process.env.NODE_ENV,
-      requestId,
-    }), {
-      status: 500,
-      headers: { 'Content-Type': 'application/json' },
-    });
+    return createErrorResponse(requestId, error);
   }
 }
